pref:迁移项目数据逻辑优化 #12117
