+++ src/backend/ci/core/misc/api-misc/src/main/kotlin/com/tencent/devops/misc/pojo/constant/MiscMessageCode.kt
@@ -52,4 +52,5 @@ object MiscMessageCode {
     const val ERROR_MIGRATING_PROJECT_NO_VALID_DB_ASSIGN = "2135004" // 无可用的迁移库，请确认项目数据是否已经成功迁移、迁移库配置等
     const val ERROR_MIGRATING_PIPELINE_DATA_FAIL = "2135005" // 迁移流水线[{0}]数据失败，请稍后再试
     const val ERROR_MIGRATING_PIPELINE_STATUS_INVALID = "2135006" // 存在构建未结束的流水线，请稍后再试
+    const val ERROR_MIGRATING_PROJECT_DATA_FAIL = "2135007" // 迁移项目[{0}]数据失败，请稍后再试
 }
