+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/pojo/process/PreMigrationResult.kt
@@ -1,3 +1,6 @@
 package com.tencent.devops.misc.pojo.process
 
+import io.swagger.v3.oas.annotations.media.Schema
+
+@Schema(title = "预迁移结果")
 data class PreMigrationResult(val routingRuleMap: Map<String, String>)
