+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/strategy/impl/pipeline/PipelineTriggerDetailMigrationStrategy.kt
