+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/strategy/impl/project/AuditResourceMigrationStrategy.kt
@@ -1,19 +1,18 @@
-package com.tencent.devops.misc.strategy.impl
+package com.tencent.devops.misc.strategy.impl.project
 
 import com.tencent.devops.misc.dao.process.ProcessDataMigrateDao
 import com.tencent.devops.misc.pojo.process.MigrationContext
 import com.tencent.devops.misc.strategy.MigrationStrategy
 import com.tencent.devops.misc.utils.PageMigrationUtil
 
-class TemplateMigrationStrategy(
-    private val processDataMigrateDao: ProcessDataMigrateDao
-) : MigrationStrategy {
+class AuditResourceMigrationStrategy(private val processDataMigrateDao: ProcessDataMigrateDao) : MigrationStrategy {
 
     override fun migrate(context: MigrationContext) {
+        // 迁移T_AUDIT_RESOURCE表数据
         PageMigrationUtil.migrateByPage(
             pageSize = PageMigrationUtil.MEDIUM_PAGE_SIZE,
-            fetch = { offset, limit ->
-                processDataMigrateDao.getTemplateRecords(
+            fetch = { offset, limit -> 
+                processDataMigrateDao.getAuditResourceRecords(
                     dslContext = context.dslContext,
                     projectId = context.projectId,
                     limit = limit,
@@ -21,9 +20,9 @@ class TemplateMigrationStrategy(
                 )
             },
             migrate = { records ->
-                processDataMigrateDao.migrateTemplateData(
+                processDataMigrateDao.migrateAuditResourceData(
                     migratingShardingDslContext = context.migratingShardingDslContext,
-                    templateRecords = records
+                    auditResourceRecords = records
                 )
             }
         )
