+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/strategy/impl/project/PipelineGroupMigrationStrategy.kt
@@ -1,19 +1,18 @@
-package com.tencent.devops.misc.strategy.impl
+package com.tencent.devops.misc.strategy.impl.project
 
 import com.tencent.devops.misc.dao.process.ProcessDataMigrateDao
 import com.tencent.devops.misc.pojo.process.MigrationContext
 import com.tencent.devops.misc.strategy.MigrationStrategy
 import com.tencent.devops.misc.utils.PageMigrationUtil
 
-class PipelineJobMutexGroupMigrationStrategy(
-    private val processDataMigrateDao: ProcessDataMigrateDao
-) : MigrationStrategy {
+class PipelineGroupMigrationStrategy(private val processDataMigrateDao: ProcessDataMigrateDao) : MigrationStrategy {
 
     override fun migrate(context: MigrationContext) {
+        // 迁移T_PIPELINE_GROUP表数据
         PageMigrationUtil.migrateByPage(
             pageSize = PageMigrationUtil.LONG_PAGE_SIZE,
-            fetch = { offset, limit ->
-                processDataMigrateDao.getPipelineJobMutexGroupRecords(
+            fetch = { offset, limit -> 
+                processDataMigrateDao.getPipelineGroupRecords(
                     dslContext = context.dslContext,
                     projectId = context.projectId,
                     limit = limit,
@@ -21,9 +20,9 @@ class PipelineJobMutexGroupMigrationStrategy(
                 )
             },
             migrate = { records ->
-                processDataMigrateDao.migratePipelineJobMutexGroupData(
+                processDataMigrateDao.migratePipelineGroupData(
                     migratingShardingDslContext = context.migratingShardingDslContext,
-                    pipelineJobMutexGroupRecords = records
+                    pipelineGroupRecords = records
                 )
             }
         )
