+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/strategy/impl/project/PipelineLabelMigrationStrategy.kt
@@ -1,17 +1,20 @@
-package com.tencent.devops.misc.strategy.impl
+package com.tencent.devops.misc.strategy.impl.project
 
 import com.tencent.devops.misc.dao.process.ProcessDataMigrateDao
 import com.tencent.devops.misc.pojo.process.MigrationContext
 import com.tencent.devops.misc.strategy.MigrationStrategy
 import com.tencent.devops.misc.utils.PageMigrationUtil
 
-class PipelineGroupMigrationStrategy(private val processDataMigrateDao: ProcessDataMigrateDao) : MigrationStrategy {
+class PipelineLabelMigrationStrategy(
+    private val processDataMigrateDao: ProcessDataMigrateDao
+) : MigrationStrategy {
 
     override fun migrate(context: MigrationContext) {
+        // 迁移T_PIPELINE_LABEL表数据
         PageMigrationUtil.migrateByPage(
             pageSize = PageMigrationUtil.LONG_PAGE_SIZE,
-            fetch = { offset, limit -> 
-                processDataMigrateDao.getPipelineGroupRecords(
+            fetch = { offset, limit ->
+                processDataMigrateDao.getPipelineLabelRecords(
                     dslContext = context.dslContext,
                     projectId = context.projectId,
                     limit = limit,
@@ -19,9 +22,9 @@ class PipelineGroupMigrationStrategy(private val processDataMigrateDao: ProcessD
                 )
             },
             migrate = { records ->
-                processDataMigrateDao.migratePipelineGroupData(
+                processDataMigrateDao.migratePipelineLabelData(
                     migratingShardingDslContext = context.migratingShardingDslContext,
-                    pipelineGroupRecords = records
+                    pipelineLabelRecords = records
                 )
             }
         )
