+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/strategy/impl/project/PipelineViewUserLastViewMigrationStrategy.kt
@@ -1,19 +1,20 @@
-package com.tencent.devops.misc.strategy.impl
+package com.tencent.devops.misc.strategy.impl.project
 
 import com.tencent.devops.misc.dao.process.ProcessDataMigrateDao
 import com.tencent.devops.misc.pojo.process.MigrationContext
 import com.tencent.devops.misc.strategy.MigrationStrategy
 import com.tencent.devops.misc.utils.PageMigrationUtil
 
-class PipelineViewTopMigrationStrategy(
+class PipelineViewUserLastViewMigrationStrategy(
     private val processDataMigrateDao: ProcessDataMigrateDao
 ) : MigrationStrategy {
 
     override fun migrate(context: MigrationContext) {
+        // 迁移T_PIPELINE_VIEW_USER_LAST_VIEW表数据
         PageMigrationUtil.migrateByPage(
             pageSize = PageMigrationUtil.LONG_PAGE_SIZE,
             fetch = { offset, limit ->
-                processDataMigrateDao.getPipelineViewTopRecords(
+                processDataMigrateDao.getPipelineViewUserLastViewRecords(
                     dslContext = context.dslContext,
                     projectId = context.projectId,
                     limit = limit,
@@ -21,9 +22,9 @@ class PipelineViewTopMigrationStrategy(
                 )
             },
             migrate = { records ->
-                processDataMigrateDao.migratePipelineViewTopData(
+                processDataMigrateDao.migratePipelineViewUserLastViewData(
                     migratingShardingDslContext = context.migratingShardingDslContext,
-                    pipelineViewTopRecords = records
+                    pipelineViewUserLastViewRecords = records
                 )
             }
         )
