+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/strategy/impl/project/ProjectPipelineTriggerEventMigrationStrategy.kt
@@ -1,19 +1,20 @@
-package com.tencent.devops.misc.strategy.impl
+package com.tencent.devops.misc.strategy.impl.project
 
 import com.tencent.devops.misc.dao.process.ProcessDataMigrateDao
 import com.tencent.devops.misc.pojo.process.MigrationContext
 import com.tencent.devops.misc.strategy.MigrationStrategy
 import com.tencent.devops.misc.utils.PageMigrationUtil
 
-class ProjectPipelineYamlViewMigrationStrategy(
+class ProjectPipelineTriggerEventMigrationStrategy(
     private val processDataMigrateDao: ProcessDataMigrateDao
 ) : MigrationStrategy {
 
     override fun migrate(context: MigrationContext) {
+        // 迁移T_PIPELINE_TRIGGER_EVENT表数据
         PageMigrationUtil.migrateByPage(
             pageSize = PageMigrationUtil.MEDIUM_PAGE_SIZE,
             fetch = { offset, limit ->
-                processDataMigrateDao.getProjectPipelineYamlViewRecords(
+                processDataMigrateDao.getProjectPipelineTriggerEventRecords(
                     dslContext = context.dslContext,
                     projectId = context.projectId,
                     limit = limit,
@@ -21,9 +22,9 @@ class ProjectPipelineYamlViewMigrationStrategy(
                 )
             },
             migrate = { records ->
-                processDataMigrateDao.migrateProjectPipelineYamlViewData(
+                processDataMigrateDao.migrateProjectPipelineTriggerEventData(
                     migratingShardingDslContext = context.migratingShardingDslContext,
-                    pipelineYamlViewRecords = records
+                    pipelineTriggerEventRecords = records
                 )
             }
         )
