+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/strategy/impl/project/ProjectPipelineYamlSyncMigrationStrategy.kt
@@ -1,19 +1,20 @@
-package com.tencent.devops.misc.strategy.impl
+package com.tencent.devops.misc.strategy.impl.project
 
 import com.tencent.devops.misc.dao.process.ProcessDataMigrateDao
 import com.tencent.devops.misc.pojo.process.MigrationContext
 import com.tencent.devops.misc.strategy.MigrationStrategy
 import com.tencent.devops.misc.utils.PageMigrationUtil
 
-class ProjectPipelineYamlBranchFileMigrationStrategy(
+class ProjectPipelineYamlSyncMigrationStrategy(
     private val processDataMigrateDao: ProcessDataMigrateDao
 ) : MigrationStrategy {
 
     override fun migrate(context: MigrationContext) {
+        // 迁移T_PIPELINE_YAML_SYNC表数据
         PageMigrationUtil.migrateByPage(
             pageSize = PageMigrationUtil.MEDIUM_PAGE_SIZE,
             fetch = { offset, limit ->
-                processDataMigrateDao.getProjectPipelineYamlBranchFileRecords(
+                processDataMigrateDao.getProjectPipelineYamlSyncRecords(
                     dslContext = context.dslContext,
                     projectId = context.projectId,
                     limit = limit,
@@ -21,9 +22,9 @@ class ProjectPipelineYamlBranchFileMigrationStrategy(
                 )
             },
             migrate = { records ->
-                processDataMigrateDao.migrateProjectPipelineYamlBranchFileData(
+                processDataMigrateDao.migrateProjectPipelineYamlSyncData(
                     migratingShardingDslContext = context.migratingShardingDslContext,
-                    pipelineYamlBranchFileRecords = records
+                    pipelineYamlSyncRecords = records
                 )
             }
         )
