+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/task/PipelineMigrationTask.kt
@@ -8,7 +8,6 @@ import com.tencent.devops.common.service.utils.SpringContextUtil
 import com.tencent.devops.common.web.utils.BkApiUtil
 import com.tencent.devops.common.web.utils.I18nUtil
 import com.tencent.devops.misc.dao.process.ProcessDao
-import com.tencent.devops.misc.factory.MigrationStrategyFactory
 import com.tencent.devops.misc.pojo.constant.MiscMessageCode
 import com.tencent.devops.misc.pojo.process.MigratePipelineDataParam
 import com.tencent.devops.misc.pojo.process.MigrationContext
@@ -28,66 +27,70 @@ class PipelineMigrationTask constructor(
         private const val RETRY_NUM = 3
         private const val DEFAULT_THREAD_SLEEP_TINE = 5000L
     }
-        override fun run() {
-            val semaphore = migratePipelineDataParam.semaphore
-            val projectId = migratePipelineDataParam.projectId
-            val pipelineId = migratePipelineDataParam.pipelineId
-            val cancelFlag = migratePipelineDataParam.cancelFlag
-            val doneSignal = migratePipelineDataParam.doneSignal
-            val dslContext = migratePipelineDataParam.dslContext
-            val migratingShardingDslContext = migratePipelineDataParam.migratingShardingDslContext
-            val migrationStrategyFactory = migratePipelineDataParam.migrationStrategyFactory
-            val processDao = migratePipelineDataParam.processDao
-            val archiveFlag = migratePipelineDataParam.archiveFlag
-            // 1、获取是否允许执行的信号量
-            semaphore?.acquire()
-            logger.info("migrateProjectData project[$projectId],pipeline[$pipelineId] start..............")
-            if (cancelFlag) {
-                // 2、取消未结束的构建
-                handleUnFinishPipelines(RETRY_NUM)
-                Thread.sleep(DEFAULT_THREAD_SLEEP_TINE)
-            }
-            // 检查构建是否结束
-            isBuildCompleted(
-                dslContext = dslContext,
-                processDao = processDao,
-                projectId = projectId,
-                pipelineId = pipelineId
-            )
-            try {
-                // 3、开始迁移流水线的数据
-                executeMigration(
-                    dslContext = dslContext,
-                    migratingShardingDslContext = migratingShardingDslContext,
-                    migrationStrategyFactory = migrationStrategyFactory,
-                    projectId = projectId,
-                    pipelineId = pipelineId,
-                    archiveFlag = archiveFlag
-                )
-                // 4、业务逻辑成功执行完后计数器减1
-                doneSignal?.countDown()
-                logger.info("migrateProjectData project[$projectId],pipeline[$pipelineId] end..............")
-            } catch (ignored: Throwable) {
-                logger.info("migrateProjectData project[$projectId],pipeline[$pipelineId] run task fail", ignored)
-                throw ErrorCodeException(
-                    errorCode = MiscMessageCode.ERROR_MIGRATING_PIPELINE_DATA_FAIL,
-                    params = arrayOf(pipelineId),
-                    defaultMessage = I18nUtil.getCodeLanMessage(
-                        messageCode = MiscMessageCode.ERROR_MIGRATING_PIPELINE_DATA_FAIL,
-                        params = arrayOf(pipelineId)
-                    )
-                )
-            } finally {
-                // 5、业务逻辑执行完成后释放信号量
-                semaphore?.release()
-            }
-        }
 
+    /**
+     * 任务主入口
+     * 1. 获取信号量许可
+     * 2. 取消未完成构建（如果配置要求）
+     * 3. 验证构建状态
+     * 4. 执行数据迁移
+     * 5. 释放资源
+     */
+    override fun run() {
+        val semaphore = migratePipelineDataParam.semaphore
+        val projectId = migratePipelineDataParam.projectId
+        val pipelineId = migratePipelineDataParam.pipelineId
+        val cancelFlag = migratePipelineDataParam.cancelFlag
+        val doneSignal = migratePipelineDataParam.doneSignal
+        val dslContext = migratePipelineDataParam.dslContext
+        val processDao = migratePipelineDataParam.processDao
+        // 1、获取是否允许执行的信号量
+        semaphore?.acquire()
+        logger.info("migrateProjectData project[$projectId],pipeline[$pipelineId] start..............")
+        if (cancelFlag) {
+            // 2、取消未结束的构建
+            handleUnFinishPipelines(RETRY_NUM)
+            Thread.sleep(DEFAULT_THREAD_SLEEP_TINE)
+        }
+        // 3、检查构建是否结束
+        isBuildCompleted(
+            dslContext = dslContext,
+            processDao = processDao,
+            projectId = projectId,
+            pipelineId = pipelineId
+        )
+        try {
+            // 4、 执行数据迁移
+            executeMigration()
+            // 5、业务逻辑成功执行完后计数器减1
+            doneSignal?.countDown()
+            logger.info("migrateProjectData project[$projectId],pipeline[$pipelineId] end..............")
+        } catch (ignored: Throwable) {
+            logger.info("migrateProjectData project[$projectId],pipeline[$pipelineId] run task fail", ignored)
+            throw ErrorCodeException(
+                errorCode = MiscMessageCode.ERROR_MIGRATING_PIPELINE_DATA_FAIL,
+                params = arrayOf(pipelineId),
+                defaultMessage = I18nUtil.getCodeLanMessage(
+                    messageCode = MiscMessageCode.ERROR_MIGRATING_PIPELINE_DATA_FAIL,
+                    params = arrayOf(pipelineId)
+                )
+            )
+        } finally {
+            // 5、业务逻辑执行完成后释放信号量
+            semaphore?.release()
+        }
+    }
+
+    /**
+     * 处理未完成的流水线构建
+     * 递归尝试取消所有运行中的构建
+     * @param retryNum 重试次数
+     */
     private fun handleUnFinishPipelines(retryNum: Int) {
         if (retryNum < 1) {
             return
         }
-        // 查看项目下是否还有未结束的构建
+        // 未完成的构建状态列表
         val unFinishStatusList = listOf(
             BuildStatus.QUEUE,
             BuildStatus.QUEUE_CACHE,
@@ -108,6 +111,7 @@ class PipelineMigrationTask constructor(
                 limit = DEFAULT_PAGE_SIZE,
                 statusList = unFinishStatusList
             )
+            // 遍历并尝试取消每个构建
             historyInfoRecords?.forEach { historyInfoRecord ->
                 val buildId = historyInfoRecord[tPipelineBuildHistory.BUILD_ID]
                 val channel = historyInfoRecord[tPipelineBuildHistory.CHANNEL]
@@ -132,6 +136,10 @@ class PipelineMigrationTask constructor(
         }
     }
 
+    /**
+     * 取消单个流水线构建
+     * @return Boolean 是否取消成功
+     */
     private fun cancelBuild(
         startUser: String,
         projectId: String,
@@ -211,22 +219,24 @@ class PipelineMigrationTask constructor(
         }
     }
 
-    private fun executeMigration(
-        dslContext: DSLContext,
-        migratingShardingDslContext: DSLContext,
-        migrationStrategyFactory: MigrationStrategyFactory,
-        projectId: String,
-        pipelineId: String,
-        archiveFlag: Boolean? = null
-    ) {
+    /**
+     * 迁移流水线级数据
+     * 使用策略模式执行不同数据表的迁移
+     */
+    private fun executeMigration() {
+        val projectId = migratePipelineDataParam.projectId
+        val pipelineId = migratePipelineDataParam.pipelineId
+        val archiveFlag = migratePipelineDataParam.archiveFlag
         val migrationContext = MigrationContext(
-            dslContext = dslContext,
-            migratingShardingDslContext = migratingShardingDslContext,
+            dslContext = migratePipelineDataParam.dslContext,
+            migratingShardingDslContext = migratePipelineDataParam.migratingShardingDslContext,
             projectId = projectId,
             pipelineId = pipelineId,
-            archiveFlag = archiveFlag
+            archiveFlag = migratePipelineDataParam.archiveFlag,
+            sourceId = if (migratePipelineDataParam.doneSignal != null) projectId else pipelineId
         )
-
+        val migrationStrategyFactory = migratePipelineDataParam.migrationStrategyFactory
+        // 获取并执行所有流水线数据迁移策略
         migrationStrategyFactory.getPipelineDataMigrationStrategies(archiveFlag).forEach { strategy ->
             strategy.migrate(migrationContext)
         }
