+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/utils/MigrationExecutor.kt
@@ -15,6 +15,7 @@ import com.tencent.devops.common.notify.enums.NotifyType
 import com.tencent.devops.common.service.utils.BkServiceUtil
 import com.tencent.devops.common.service.utils.CommonUtils
 import com.tencent.devops.common.web.utils.BkApiUtil
+import com.tencent.devops.common.web.utils.I18nUtil
 import com.tencent.devops.misc.lock.MigrationLock
 import com.tencent.devops.misc.pojo.constant.MiscMessageCode
 import com.tencent.devops.misc.pojo.process.DeleteDataParam
@@ -48,30 +49,53 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         private const val MIGRATE_PROCESS_PROJECT_DATA_PROJECT_COUNT_KEY = "MIGRATE_PROCESS_PROJECT_DATA_PROJECT_COUNT"
     }
 
+    /**
+     * 迁移执行主入口
+     * 1. 准备迁移环境
+     * 2. 执行数据迁移
+     * 3. 完成迁移后处理
+     * 4. 异常处理和资源清理
+     */
     fun execute() {
         val projectId = config.projectId
         logger.info("Starting migration for project $projectId")
         try {
+            // 1. 准备迁移环境
             prepareMigration()
+            // 2. 执行数据迁移
             migrateProjectData()
+            // 3. 完成迁移后处理
             completeMigration()
             logger.info("Migration completed successfully for project $projectId")
         } catch (ignored: Throwable) {
+            // 迁移失败处理
             handleMigrationFailure(ignored)
             throw ignored
         } finally {
+            // 资源清理
             cleanupResources()
         }
     }
 
+    /**
+     * 准备迁移环境
+     * 1. 将项目标记为迁移中状态
+     * 2. 清理目标数据库中的旧数据
+     */
     private fun prepareMigration() {
+        // 在Redis中标记项目为迁移中状态
         config.redisOperation.addSetValue(
             key = MiscUtils.getMigratingProjectsRedisKey(SystemModuleEnum.PROCESS.name),
             item = config.projectId
         )
+        // 删除目标数据库中的旧数据
         deleteTargetData()
     }
 
+    /**
+     * 删除目标数据库中的现有数据
+     * 防止重复数据导致迁移失败
+     */
     private fun deleteTargetData() {
         val targetDataSourceName = getTargetDataSourceName()
         val deleteDataParam = DeleteDataParam(
@@ -84,33 +108,54 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         config.processDataDeleteService.deleteProcessData(deleteDataParam)
     }
 
+    /**
+     * 项目数据迁移
+     * 1. 先迁移流水线数据（多线程并行）
+     * 2. 再迁移项目级数据（单线程顺序执行）
+     */
     private fun migrateProjectData() {
-        migrateProjectDirectData()
+        // 先启动耗时的多线程迁移流水线维度数据
         migratePipelineData()
+        // 然后执行单线程迁移项目维度数据
+        migrateProjectDirectData()
     }
 
+    /**
+     * 迁移项目级数据
+     * 使用策略模式执行不同数据表的迁移
+     */
     private fun migrateProjectDirectData() {
         val migrationContext = MigrationContext(
             dslContext = config.dslContext,
             migratingShardingDslContext = config.migratingShardingDslContext,
             projectId = config.projectId
         )
+        // 遍历所有项目级数据迁移策略并执行
         config.migrationStrategyFactory.getProjectDataMigrationStrategies().forEach { strategy ->
             strategy.migrate(migrationContext)
         }
     }
 
+    /**
+     * 迁移流水线维度数据
+     * 1. 分页查询流水线ID列表
+     * 2. 使用线程池并行迁移每个流水线
+     * 3. 使用CountDownLatch等待所有任务完成
+     * 4. 添加超时机制防止永久阻塞
+     */
     private fun migratePipelineData() {
         val projectId= config.projectId
+        // 获取项目下流水线总数
         val pipelineNum = config.processDao.getPipelineNumByProjectId(config.dslContext, projectId)
+        // 无流水线则终止流程
         if (pipelineNum <= 0) return
-
+        // 动态计算线程池大小（不超过默认最大值）
         val threadNum = if (pipelineNum < DEFAULT_THREAD_NUM) {
             pipelineNum
         } else {
             DEFAULT_THREAD_NUM
         }
-
+        // 创建线程池和同步工具
         val executor = Executors.newFixedThreadPool(threadNum)
         val semaphore = Semaphore(threadNum)
         val doneSignal = CountDownLatch(pipelineNum)
@@ -119,6 +164,7 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         try {
             var minPipelineInfoId = processDao.getMinPipelineInfoIdByProjectId(dslContext, projectId)
             do {
+                // 分页查询流水线ID列表
                 val pipelineIdList = processDao.getPipelineIdListByProjectId(
                     dslContext = dslContext,
                     projectId = projectId,
@@ -126,6 +172,7 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
                     limit = DEFAULT_PAGE_SIZE.toLong()
                 )?.map { it.getValue(0).toString() }
 
+                // 提交每个流水线的迁移任务
                 pipelineIdList?.forEach { pipelineId ->
                     executor.submit(
                         PipelineMigrationTask(
@@ -142,17 +189,33 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
                             )
                         )
                     )
-                    minPipelineInfoId = updateMinPipelineId(pipelineIdList, minPipelineInfoId)
+                }
+                // 更新分页起始ID（基于最后一条记录的ID）
+                if (!pipelineIdList.isNullOrEmpty()) {
+                    minPipelineInfoId = (processDao.getPipelineInfoByPipelineId(
+                        dslContext = dslContext,
+                        projectId = projectId,
+                        pipelineId = pipelineIdList.last()
+                    )?.id ?: 0L) + 1
                 }
             } while (pipelineIdList?.size == DEFAULT_PAGE_SIZE)
 
-            // 添加超时机制防止永久阻塞
+            // 等待所有任务完成（带超时机制）
             if (!doneSignal.await(config.migrationTimeout, TimeUnit.HOURS)) {
                 logger.error("Pipeline migration timed out for project ${config.projectId}")
+                throw ErrorCodeException(
+                    errorCode = MiscMessageCode.ERROR_MIGRATING_PROJECT_DATA_FAIL,
+                    params = arrayOf(projectId),
+                    defaultMessage = I18nUtil.getCodeLanMessage(
+                        messageCode = MiscMessageCode.ERROR_MIGRATING_PROJECT_DATA_FAIL,
+                        params = arrayOf(projectId)
+                    )
+                )
             }
         } finally {
             // 确保资源释放
             executor.shutdownNow()
+            // 释放信号量资源
             releaseSemaphorePermits(semaphore, threadNum)
         }
     }
@@ -163,25 +226,30 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         }
     }
 
-    private fun updateMinPipelineId(pipelineIdList: List<String>, currentMinId: Long): Long {
-        return pipelineIdList.lastOrNull()?.let { lastPipelineId ->
-            (config.processDao.getPipelineInfoByPipelineId(
-                dslContext = config.dslContext,
-                projectId = config.projectId,
-                pipelineId = lastPipelineId
-            )?.id ?: 0L) + 1
-        } ?: currentMinId
-    }
-
+    /**
+     * 迁移完成处理：
+     * 1. 选择性删除源数据（配置决定）
+     * 2. 更新分片路由规则
+     * 3. 保存迁移历史记录
+     * 4. 发送成功通知
+     */
     private fun completeMigration() {
+        // 根据配置决定是否删除源数据库数据
         if (config.migrationSourceDbDataDeleteFlag) {
             deleteSourceData()
         }
+        // 更新路由规则
         updateShardingRules()
+        // 保存迁移历史记录
         saveMigrationHistory()
+        // 发送迁移成功通知
         sendSuccessNotification()
     }
 
+    /**
+     * 删除源数据库中的数据
+     * （仅在配置开启时执行）
+     */
     private fun deleteSourceData() {
         val targetDataSourceName = getTargetDataSourceName()
         val deleteDataParam = DeleteDataParam(
@@ -194,6 +262,12 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         config.processDataDeleteService.deleteProcessData(deleteDataParam)
     }
 
+    /**
+     * 更新分片路由规则：
+     * 1. 将项目路由指向新数据源
+     * 2. 清理迁移临时路由
+     * 3. 等待并验证所有微服务缓存更新
+     */
     private fun updateShardingRules() {
         val projectId = config.projectId
         val migratingShardingRoutingRuleKey = ShardingUtil.getMigratingShardingRoutingRuleKey(
@@ -215,7 +289,7 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
                 routingRule = targetDataSourceName
             )
         )
-        // 清除缓存中项目的迁移DB路由规则
+        // 清理临时路由规则
         config.redisOperation.delete(migratingShardingRoutingRuleKey)
         // 睡眠一会儿等待服务器缓存更新
         Thread.sleep(DEFAULT_THREAD_SLEEP_TIMEOUT)
@@ -223,10 +297,14 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         confirmAllServiceCacheUpdated()
     }
 
+    /**
+     * 确认所有微服务的路由缓存已更新
+     * 循环检查所有服务的实例是否完成更新
+     */
     private fun confirmAllServiceCacheUpdated() {
         val serviceNames = config.migrationProcessDbMicroServices.split(",")
         serviceNames.forEach { serviceName ->
-            logger.info("service[$serviceName] confirmCacheIsUpdated start")
+            logger.info("Verifying cache update for service [$serviceName]")
             val cacheKey = ShardingUtil.getShardingRoutingRuleKey(
                 clusterName = clusterName,
                 moduleCode = SystemModuleEnum.PROCESS.name,
@@ -244,12 +322,20 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         }
     }
 
+    /**
+     * 递归检查单个服务的缓存更新状态
+     * @param serviceName 服务名称
+     * @param cacheKey 路由缓存key
+     * @param retryNum 重试次数
+     * @return Boolean 是否所有实例都更新完成
+     */
     private fun confirmCacheUpdated(serviceName: String, cacheKey: String, retryNum: Int): Boolean {
         if (retryNum <= 0) return false
         
         val finalServiceName = BkServiceUtil.findServiceName(serviceName = serviceName)
         val serviceHostKey = BkServiceUtil.getServiceHostKey(finalServiceName)
         val redisOperation = config.redisOperation
+        // 获取该服务的所有实例IP
         val serviceIps = redisOperation.getSetMembers(serviceHostKey)?.toMutableSet() ?: mutableSetOf()
         val serviceRoutingRuleActionFinishKey = BkServiceUtil.getServiceRoutingRuleActionFinishKey(
             serviceName = finalServiceName,
@@ -262,11 +348,16 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
             "confirmCacheIsUpdated service[$serviceName] cacheKey:$cacheKey retryNum:$retryNum " +
                     "serviceIps:$serviceIps finishServiceIps:$finishServiceIps"
         )
+        // 计算未更新的实例
         serviceIps.removeAll(finishServiceIps)
         return if (serviceIps.isEmpty()) true
         else confirmCacheUpdated(serviceName, cacheKey, retryNum - 1)
     }
 
+    /**
+     * 保存迁移历史记录
+     * 记录源和目标数据源信息
+     */
     private fun saveMigrationHistory() {
         config.projectDataMigrateHistoryService.add(
             userId = config.userId,
@@ -283,6 +374,9 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         )
     }
 
+    /**
+     * 发送迁移成功通知
+     */
     private fun sendSuccessNotification() {
         val projectId = config.projectId
         val pipelineNum = config.processDao.getPipelineNumByProjectId(config.dslContext, projectId)
@@ -305,16 +399,31 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         }
     }
 
+    /**
+     * 迁移失败处理：
+     * 1. 记录错误日志
+     * 2. 执行回滚操作
+     * 3. 发送失败通知
+     */
     private fun handleMigrationFailure(ignored: Throwable) {
-        logger.warn("Migration failed for project ${config.projectId}", ignored)
+        logger.error("Migration failed for project ${config.projectId}", ignored)
+        // 回滚已迁移的数据
         rollbackMigration()
+        // 发送迁移失败通知
         sendFailureNotification(ignored.message)
     }
 
+    /**
+     * 回滚操作：
+     * 1. 删除目标数据库中的迁移数据
+     * 2. 恢复原始路由规则
+     * 3. 验证服务缓存回滚
+     */
     private fun rollbackMigration() {
         val projectId = config.projectId
         try {
             val targetDataSourceName = getTargetDataSourceName()
+            // 删除目标数据库中的迁移数据
             val deleteDataParam = DeleteDataParam(
                 dslContext = config.migratingShardingDslContext,
                 projectId = projectId,
@@ -324,6 +433,7 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
             )
             config.processDataDeleteService.deleteProcessData(deleteDataParam)
             val client = config.client
+            // 恢复原始路由规则
             val historyShardingRoutingRule = client.get(ServiceShardingRoutingRuleResource::class)
                 .getShardingRoutingRuleByName(
                     routingName = projectId,
@@ -342,13 +452,16 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
                 userId = config.userId,
                 shardingRoutingRule = historyShardingRoutingRule
             )
-
+            // 验证服务缓存是否已回滚
             confirmAllServiceCacheUpdated()
         } catch (ignored: Throwable) {
             logger.warn("Rollback failed for project $projectId", ignored)
         }
     }
 
+    /**
+     * 发送迁移失败通知
+     */
     private fun sendFailureNotification(errorMsg: String?) {
         val projectId = config.projectId
         val titleParams = mapOf(KEY_PROJECT_ID to projectId)
@@ -371,21 +484,32 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         }
     }
 
+    /**
+     * 资源清理：
+     * 1. 移除迁移中状态
+     * 2. 更新并发计数
+     * 3. 解除项目锁定
+     * 4. 清理执行计数
+     */
     private fun cleanupResources() {
         val redisOperation = config.redisOperation
         val projectId = config.projectId
+        // 从迁移中项目集合移除该项目
         redisOperation.removeSetMember(
             key = MiscUtils.getMigratingProjectsRedisKey(SystemModuleEnum.PROCESS.name),
             item = projectId
         )
+        // 减少全局迁移项目计数
         redisOperation.increment(
             key = MIGRATE_PROCESS_PROJECT_DATA_PROJECT_COUNT_KEY,
             incr = -1
         )
+        // 解除项目的API访问限制
         redisOperation.removeSetMember(
             key = BkApiUtil.getApiAccessLimitProjectsKey(),
             item = projectId
         )
+        // 清理项目执行计数
         redisOperation.delete(getMigrateProjectExecuteCountKey(projectId))
     }
 
@@ -393,6 +517,10 @@ class MigrationExecutor(private val config: MigrationExecutorConfig) {
         return "MIGRATE_PROJECT_PROCESS_DATA_EXECUTE_COUNT:$projectId"
     }
 
+    /**
+     * 获取目标数据源名称
+     * 从预迁移结果中提取路由规则映射
+     */
     private fun getTargetDataSourceName(): String {
         val preMigrationResult = config.preMigrationResult
         val routingRule = preMigrationResult.routingRuleMap.keys.firstOrNull()
