+++ support-files/i18n/misc/message_en_US.properties
@@ -1,10 +1,11 @@
 # misc
 2135001=The number of projects migrated at the same time has exceeded {0}, please try again later
 2135002=Update microservice [{0}] server sharding rule local cache failed
-2135003=The data of project [{0}] has been migrated and no further migration is required
+2135003=The data of project [{0}] is being migrated or has been migrated. There is no need to migrate again
 2135004=There is no migration library available. Please confirm whether the project data has been successfully migrated, the migration library configuration, etc
 2135005=Failed to migrate pipeline [{0}] data, please try again later
 2135006=There are pipelines that have not been completed yet. Please try again later
+2135007=Failed to migrate project [{0}] data, please try again later
 # image
 2107001=User ({0}) does not have permission to upload image
 2107002=The number of valid images for image files is zero.
