+++ support-files/i18n/misc/message_ja_JP.properties
@@ -1,11 +1,11 @@
 #misc
 2135001=同時に移行するプロジェクト数が{0}個を超えています。しばらくしてから再試行してください
 2135002=マイクロサービス[{0}]のサーバーシャーディングルールのローカルキャッシュの更新に失敗しました
-2135003=プロジェクト[{0}]のデータは既に移行済みです。再度移行する必要はありません
+2135003=プロジェクト [{0}] のデータは移行中または移行済みです。再度移行する必要はありません
 2135004=利用可能な移行リポジトリがありません。プロジェクトデータが正常に移行されたか、移行リポジトリの設定などを確認してください
 2135005=パイプライン[{0}]のデータ移行に失敗しました。しばらくしてから再試行してください
 2135006=ビルドが完了していないパイプラインが存在します。しばらくしてから再試行してください
-2135007=PACモードが有効化されたパイプライン[{0}]をアーカイブするには、デフォルトブランチ上のyamlファイルを削除する必要があります
+2135007=プロジェクト [{0}] のデータを移行できませんでした。しばらくしてからもう一度お試しください
 #image
 2107001=ユーザー({0})にイメージアップロード権限がありません
 2107002=イメージファイルに有効なイメージが含まれていません
