+++ support-files/i18n/misc/message_zh_CN.properties
@@ -1,10 +1,11 @@
 # misc
 2135001=同时迁移的项目数量已超过{0}个，请稍后再试
 2135002=更新微服务[{0}]服务器分片规则本地缓存失败
-2135003=项目[{0}]的数据已经迁移完成，无需再迁移
+2135003=项目[{0}]的数据正在迁移或者已经迁移完成，无需重复迁移
 2135004=无可用的迁移库，请确认项目数据是否已经成功迁移、迁移库配置等
 2135005=迁移流水线[{0}]数据失败，请稍后再试
 2135006=存在构建未结束的流水线，请稍后再试
+2135007=迁移项目[{0}]数据失败，请稍后再试
 # image
 2107001=用户({0})没有上传镜像权限
 2107002=镜像文件有效镜像个数为零
